requirements.txt
# Core
numpy>=1.24,<3.0
pandas>=2.0,<3.0
scipy>=1.10,<2.0
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
joblib>=1.3,<2.0
PyYAML>=6.0.1,<7.0

# Optional (uncomment if you use them)
# lightgbm>=4.0,<5.0
# catboost>=1.2,<2.0
# matplotlib>=3.8,<4.0